Refactor settings page into nested routes with sidebar
Splits the 1600-line monolithic settings.tsx into 14 card components
under web/src/components/settings/, plus 6 category routes under
/settings/{system,safety,account,appearance,data,danger}. /settings
redirects to /settings/system; existing #safety / #security deep
links and the safety/retention banners now point at concrete sub-paths.

Onboarding popup gains an "Import existing data" row pointing at /import.
