+++ web/src/lib/utils.ts
@@ -23,6 +23,14 @@ export function formatDateTime(iso: string): string {
   })
 }
 
+export function formatBytes(bytes: number): string {
+  if (bytes === 0) return "0 B"
+  const units = ["B", "KB", "MB", "GB"]
+  const i = Math.floor(Math.log(bytes) / Math.log(1024))
+  const value = bytes / Math.pow(1024, i)
+  return `${value.toFixed(i > 0 ? 1 : 0)} ${units[i]}`
+}
+
 export function timeAgo(iso: string): string {
   const seconds = Math.floor((Date.now() - new Date(iso).getTime()) / 1000)
   if (seconds < 60) return "just now"
