+++ web/src/routes/settings/danger.tsx
@@ -1,0 +1,9 @@
+import { DeleteAccountCard } from "@/components/settings/delete-account-card";
+
+export function SettingsDangerPage() {
+  return (
+    <>
+      <DeleteAccountCard />
+    </>
+  );
+}
