+++ web/src/routes/settings/index.tsx
@@ -1,0 +1,5 @@
+import { Navigate } from "react-router";
+
+export function SettingsIndex() {
+  return <Navigate to="/settings/system" replace />;
+}
