+++ web/src/routes/settings/safety.tsx
@@ -1,0 +1,11 @@
+import { SystemSafetyCard } from "@/components/system-safety-card";
+import { RevisionRetentionCard } from "@/components/revision-retention-card";
+
+export function SettingsSafetyPage() {
+  return (
+    <>
+      <SystemSafetyCard />
+      <RevisionRetentionCard />
+    </>
+  );
+}
